Initial changes for proect detail files and versions
