+++ web-app/packages/admin-lib/src/modules/admin/views/index.ts
@@ -8,3 +8,4 @@ export { default as SettingsView } from './SettingsView.vue'
 export { default as SettingsViewTemplate } from './SettingsViewTemplate.vue'
 export { default as ProjectsView } from './ProjectsView.vue'
 export { default as ProjectView } from './ProjectView.vue'
+export { default as ProjectSettingsView } from './ProjectSettingsView.vue'
