+++ web-app/packages/lib/src/common/route_utils.ts
@@ -45,16 +45,27 @@ export function isAuthenticatedGuard(
   }
 }
 
-/** Handles redirect to /login when user is not superUser. */
+/** Handles redirect to /login when user is not superUser and authenticated.
+ *
+ * Usage in admin routes
+*/
 export function isSuperUser(
   to: RouteLocationNormalized,
   from: RouteLocationNormalized,
   next: NavigationGuardNext,
-  userStore
+  userStore,
+  options?: IsAuthenticatedGuardOptions
 ) {
-  if (userStore.isSuperUser) {
-    next()
+  if (to.meta.public || (userStore.isLoggedIn && userStore.isSuperUser)) {
+    if (isTheSameRoute(from, to)) {
+      return false
+    } else {
+      next()
+    }
   } else {
-    next('/login')
+    next({
+      path: options?.notAuthenticatedRedirectPath ?? 'login',
+      query: { redirect: to.fullPath }
+    })
   }
 }
