+++ web-app/packages/lib/src/mm-theme.ts
@@ -11,6 +11,7 @@ import { MenuPassThroughOptions } from 'primevue/menu'
 import { usePassThrough } from 'primevue/passthrough'
 import { PasswordPassThroughOptions } from 'primevue/password'
 import { ProgressBarPassThroughOptions } from 'primevue/progressbar'
+import { TabPanelPassThroughOptions } from 'primevue/tabpanel'
 import { TagPassThroughOptions } from 'primevue/tag'
 import { ToastPassThroughOptions } from 'primevue/toast'
 import { TreePassThroughOptions } from 'primevue/tree'
@@ -182,7 +183,24 @@ export default usePassThrough(
           class: 'border-noround border-x-none'
         }
       }
-    } as AccordionPassThroughOptions
+    } as AccordionPassThroughOptions,
+    tabPanel: {
+      headerAction({ context }) {
+        // Custom handling of active styles for tabs
+        return {
+          style: {
+            backgroundColor: 'transparent',
+            borderBottomColor: context?.active
+              ? 'var(--forest-color)'
+              : 'transparent'
+          },
+          class: [
+            'hover:border-400 pb-4',
+            { 'text-color-forest': context?.active }
+          ]
+        }
+      }
+    } as TabPanelPassThroughOptions
   },
   {}
 )
