+++ web-app/packages/lib/src/modules/project/views/ProjectViewTemplate.vue
@@ -63,13 +63,12 @@ SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-MerginMaps-Commercial
           }
         }"
       >
-        <PTabPanel :header="tabs[0].header" :pt="ptHeaderAction"></PTabPanel>
-        <slot name="map.tab" :ptHeaderAction="ptHeaderAction" />
+        <PTabPanel :header="tabs[0].header"></PTabPanel>
+        <slot name="map.tab" />
         <!-- Render other tabs with header -->
         <PTabPanel
           v-for="tab in tabs.slice(1).filter((item) => item.header)"
           :header="tab.header"
-          :pt="ptHeaderAction"
           :key="tab.route"
         ></PTabPanel>
       </PTabView>
@@ -117,7 +116,6 @@ SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-MerginMaps-Commercial
 
 <script lang="ts">
 import { mapActions, mapState } from 'pinia'
-import { TabPanelPassThroughOptions } from 'primevue/tabpanel'
 import { defineComponent, PropType } from 'vue'
 
 import { AppContainer, AppSection } from '@/common'
@@ -212,27 +210,6 @@ export default defineComponent({
       return this.tabs.findIndex((item) =>
         this.$route.matched.some((m) => m.name === item.route)
       )
-    },
-
-    /** Rewrite of styles for TabPanels */
-    ptHeaderAction(): TabPanelPassThroughOptions {
-      return {
-        headerAction({ context }) {
-          // Custom handling of active styles for tabs
-          return {
-            style: {
-              backgroundColor: 'transparent',
-              borderBottomColor: context.active
-                ? 'var(--forest-color)'
-                : 'transparent'
-            },
-            class: [
-              'hover:border-400 pb-4',
-              { 'text-color-forest': context.active }
-            ]
-          }
-        }
-      }
     },
 
     canCloneProject() {
